Gui improvements : moving to NavigationSplitView
